Update data_processing.py
Adding age column to the data.
Age = Date of player attribute data update - Date of birth

Tracking the age of the player.
We could have just used the date of birth, but that would give us the age of player on the day we run the code. But we want the player's age on the day the data was updated(by FIFA).