Quality Analysis performed on Positions extracted
QA performed by analyzing the line up of each player for 3 different time buckets: 2008- 2010, 2011-2013, 2014-2016. The mean for each bucket is compared with the overall mean to understand player's deviation in his position through time. Since most players ( over 91%) in the data set do not have variations in their play, it is concluded that the approach to position extraction is fine.